\name{M3-package}
\alias{M3-package}
\alias{M3}

\docType{package}

\title{Functions to read in and manipulate air quality model output from
  files in Models3 format
}

\description{This package contains functions to read in and manipulate
  air quality model output from Models3-formatted files.  This format is
  used by the Community Multiscale Air Quaility (CMAQ) model. 
}


\references{For detailed information about the format and conventions of
  the Models3 format, see \cr
  \url{http://www.baronams.com/products/ioapi}.
}


\note{This software program has been reviewed in accordance with
  U.S. Environmental Protection Agency policy and approved for
  publication.  Mention of trade names or commercial products does not
  constitute endorsement or recommendation for use.
}


\keyword{IO}
