\name{M3-package}
\alias{M3-package}
\alias{M3}

\docType{package}

\title{Functions to read in and manipulate air quality model output from
  files in Models3 format
}

\description{Provides functions to read in and manipulate
  air quality model output from Models3-formatted files.  This format is
  used by the Community Multiscale Air Quaility (CMAQ) model. 
}


\references{For detailed information about the format and conventions of
  the Models3 format, see \cr
  \url{https://www.cmascenter.org/ioapi/documentation/all_versions/html/AA.html}.
}


\note{The original version of this software program was reviewed in accordance with
  U.S. Environmental Protection Agency policy and approved for publication. 
  Subsequent releases have not been reviewed by U.S. EPA, since author is no
  longer employed there.  Mention of trade names or commercial products does
  not constitute endorsement or recommendation for use.
}


\keyword{IO}
